The test us a unit test, not functional. No need for the test yaml file either.
