--- tests/pytests/functional/modules/testyaml.yaml
@@ -1,2 +1,0 @@
-key:
-  value
